pyproject.toml
[project]
name = "telegram-ai-chat-bot"
version = "0.1.0"
description = "Telegram cashback bots (sellers + clients)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [

    "requests==2.32.5",

    # config
    "pydantic==2.11.7",
    "pydantic-settings==2.12.0",


    # Google Sheets
    "gspread==6.0.2",
    "gspread-asyncio==2.0.0",
    "google-auth==2.41.1",
    "google-auth-oauthlib==1.2.2",

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "pyproject.toml", "README.md"]